Resume.csv/requirements_basic.txt
flask==2.3.3
pandas==2.0.3
PyPDF2==3.0.1
python-docx==0.8.11
openpyxl==3.1.2
werkzeug==2.3.7
flask-cors==4.0.0 